requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2
langchain
langchain_google_genai
langchain_community
faiss-cpu
sentence-transformers